6th day
took a lot of time to realize adding an obstacle AFTER passing through that point is impossible(duh). there is probably a much nicer way to write two array.folds that find the guard's position with Array.pick etc. but I am to frustrated to work on that now.
added a timing code after seeing the memes about bruteforce taking too much time
i decided that it is more accurate to say this "challenge" is not about no extra "variable declaration" but using no let statement
i think i am getting better at piping and it is hard to say this is not a variable declaration

about the code:
first find the position and the direction of the guard and pass them as a tuple to a Array.unfold
unfold basically runs until the guard gets out of bounds for ans1
for ans2, there is another unfold process to simulate having an obstacle at the point the guard will next step onto
it runs until
 1)it finds that it loops itself:
  basically (i, j) directions((0, 1), (-1, 0) etc.) turn into 2 4 8 16 using 4*(i + 3*abs(i)) + (j + 3 * abs(j))
  and the directions are map to an integer using bitmask(||| to set &&& to check)
 2)it goes out of bounds
if a loop is found, the position is saved and finally the number of possible positions are calculated